@@ -1,5 +1,5 @@
-import express = require("express");
-import repository = require("../repository/TodoRepository");
+import * as express from "express";
+import * as repository from "../repository/TodoRepository";
 
 export function GetAllTodos(req: express.Request, res: express.Response) {
 	repository.getTodos().then((todos) => {
